Added float PressurePlateThresholdMass to OpenDoor.h, this variable will track the mass required to trigger the specified pressure place. This variable is exposed to the editor.
Added float TotalMassofActors() to OpenDoor.h, this function calculates
the total mass of actors overlapping the specified pressure plate.

The door now opens based upon the mass placed in the pressure plate so
the items in the map can be placed on the pressure plate to keep the
door open and allow the player to escape.